Move create_description_table function into own module.
